migrate test_parallels to pytest
